feat(linkedin): clean
