--- logs/error_logs.json
@@ -1,10 +1,0 @@
-{
-    "sunithvs": {
-        "timestamp": "2024-12-13T07:10:49.058942",
-        "error": "Expected an expression, got 'end of print statement'"
-    },
-    "rohittp0": {
-        "timestamp": "2024-12-13T07:05:23.839656",
-        "error": "unexpected '}'"
-    }
-}
